changed value from int to double

+++ src/main/java/kernbeisser/DBEntities/Transaction.java
@@ -17,7 +17,7 @@ public class Transaction {
 
     @Column
     //TODO save as double
-    private int value;
+    private double value;
 
     @JoinColumn
     @ManyToOne
@@ -48,7 +48,7 @@ public class Transaction {
         return Tools.getAll(Transaction.class, condition);
     }
 
-    public static void doTransaction(User from, User to, int value) {
+    public static void doTransaction(User from, User to, double value) {
         EntityManager em = DBConnection.getEntityManager();
         EntityTransaction et = em.getTransaction();
         et.begin();
@@ -70,11 +70,11 @@ public class Transaction {
         em.close();
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
